+++ test/jest/helpers/intl.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import { IntlProvider } from 'react-intl';
+
+import componentsTranslations from '@folio/stripes-components/translations/stripes-components/en';
+import smartComponentsTranslations from '@folio/stripes-smart-components/translations/stripes-smart-components/en';
+import stripesCoreTranslations from '@folio/stripes-core/translations/stripes-core/en';
+import eHoldingsTranslations from '../../../translations/ui-eholdings/en.json';
+
+const prefixKeys = (translations, prefix) => {
+  return Object
+    .keys(translations)
+    .reduce((acc, key) => (
+      {
+        ...acc,
+        [`${prefix}.${key}`]: translations[key],
+      }
+    ), {});
+};
+
+const translations = {
+  ...prefixKeys(eHoldingsTranslations, 'ui-eholdings'),
+  ...prefixKeys(componentsTranslations, 'stripes-components'),
+  ...prefixKeys(smartComponentsTranslations, 'stripes-smart-components'),
+  ...prefixKeys(stripesCoreTranslations, 'stripes-core'),
+};
+
+const Intl = ({ children }) => (
+  <IntlProvider
+    locale="en"
+    messages={translations}
+  >
+    {children}
+  </IntlProvider>
+);
+
+export default Intl;